+++ CHANGELOG.md
@@ -3,6 +3,11 @@
 - Wait, I have a changelog?
 - Added poe.ninja support with proper slash commands for generic price checks. Still working through the item types. Check the readme for more info.
 - Added poe.ninja cluster jewel support
+- Added poe.ninja map support
+- Added poe.ninja blighted map support
+- Added poe.ninja blight-ravaged map support
+- Added poe.ninja unique map support
+- Added poe.ninja delirium orb support
 
 ## 1.0.0